população banco

+++ app/src/main/java/com/apphamba/hamba/titulo/dominio/Episodio.java
@@ -2,15 +2,15 @@ package com.apphamba.hamba.titulo.dominio;
 
 public class Episodio {
     private int id;
-    private int idTemoporada;
+    private long idTemoporada;
     private String nome;
     private int numeroEpisodio;
 
-    public int getIdTemoporada() {
+    public long getIdTemoporada() {
         return idTemoporada;
     }
 
-    public void setIdTemoporada(int idTemoporada) {
+    public void setIdTemoporada(long idTemoporada) {
         this.idTemoporada = idTemoporada;
     }
